t989 fixed avgNumOfAssessmentPerClinicianClinic issue from int to float and enabled isStretchWithOverflow for title to wrap

+++ escreening/src/main/java/gov/va/escreening/repository/VeteranAssessmentRepository.java
@@ -112,7 +112,7 @@ public interface VeteranAssessmentRepository extends RepositoryInterface<Veteran
 
     Integer getVetWithMultipleBatteriesFor593(String fromDate, String toDate, List<Integer> clinicIds);
 
-    Integer getAvgNumOfAssessmentPerClinicianClinicFor593(String fromDate, String toDate, List<Integer> clinicIds);
+    float getAvgNumOfAssessmentPerClinicianClinicFor593(String fromDate, String toDate, List<Integer> clinicIds);
 
     List<Report593ByDayDTO> getBatteriesByDayFor593(String fromDate, String toDate, List<Integer> clinicIds);
